# apps/api_v2/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "api_v2",
    "description": "Public markdown content, categories and search text, and employee-authenticated editor operations. Editor reads, writes, and the version token require a GitHub-authenticated govtech-bb organization member session.",
    "version": "0.0.0"
  },
  "components": {
    "securitySchemes": {
      "previewToken": {
        "type": "apiKey",
        "in": "header",
        "name": "x-preview-token",
        "description": "The site's PREVIEW_SECRET, sent server to server. Optional on site reads; with it, `preview` content is served too."
      },
      "editorSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token",
        "description": "HTTP-only employee session cookie; Secure-prefixed in production."
      }
    },
...
  },
  "paths": {
...
    "/services": {
      "get": {
        "summary": "List services",
        "tags": [
          "pages"
        ],
        "description": "The editor's index, ordered by title. A categorised page at the root of its category is an entry, and every page beneath it by `parent_id` belongs to it.",
        "security": [
          {
            "editorSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid",
                        "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12})$",
                        "description": "The entry page's id."
                      },
                      "url": {
                        "type": "string"
                      },
                      "title": {
                        "type": "string"
                      },
                      "category": {
                        "type": "object",
                        "properties": {
                          "slug": {
                            "type": "string"
                          },
                          "title": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "slug",
                          "title"
                        ],
                        "additionalProperties": false
                      },
                      "visibility": {
                        "type": "string",
                        "enum": [
                          "public",
                          "preview",
                          "draft"
                        ]
                      },
                      "form_id": {
                        "nullable": true,
                        "description": "The entry page's form, else its `/start` page's.",
                        "type": "string"
                      },
                      "has_start_page": {
                        "type": "boolean"
                      },
                      "page_count": {
                        "type": "integer",
                        "minimum": -9007199254740991,
                        "maximum": 9007199254740991,
                        "description": "The entry page plus every page below it."
                      },
                      "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
                        "description": "The latest change across those pages."
                      }
                    },
                    "required": [
                      "id",
                      "url",
                      "title",
                      "category",
                      "visibility",
                      "form_id",
                      "has_start_page",
                      "page_count",
                      "updated_at"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": {}
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": {}
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": {}
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": {}
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}